+++ db/migrate/20210216220353_add_y_to_sections.rb
@@ -1,5 +1,5 @@
 class AddYToSections < ActiveRecord::Migration[6.1]
   def change
-    add_column :sections, :y, :integer
+    add_column :sections, :y, :float
   end
 end
